+++ flexiapi/database/migrations/2024_04_08_144054_add_request_counter_and_ip_to_api_keys_table.php
@@ -1,0 +1,25 @@
+<?php
+
+use App\ApiKey;
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+    public function up()
+    {
+        Schema::table('api_keys', function (Blueprint $table) {
+            $table->string('ip')->nullable();
+            $table->integer('requests')->default(0);
+        });
+    }
+
+    public function down()
+    {
+        Schema::table('api_keys', function (Blueprint $table) {
+            $table->dropColumn('ip');
+            $table->dropColumn('requests');
+        });
+    }
+};